% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNet-methods.R
\name{print.ordinalNet}
\alias{print.ordinalNet}
\title{Print method for an "ordinalNet" object.}
\usage{
\method{print}{ordinalNet}(x, ...)
}
\arguments{
\item{x}{An "ordinalNet" S3 object}

\item{...}{Not used. Additional summary arguments.}
}
\description{
Provides a model fit summary as a data frame. For each \code{lambda} value in
the solution path, the following information is included: degrees of freedom
(number of nonzero parameters), log-likelihood, AIC, BIC, and percent deviance explained.
}
