% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNetCV.R
\name{print.ordinalNetCV}
\alias{print.ordinalNetCV}
\title{Print method for an "ordinalNetCV" object.}
\usage{
\method{print}{ordinalNetCV}(x, ...)
}
\arguments{
\item{x}{An "ordinalNetCV" S3 object}

\item{...}{Not used. Additional print arguments.}
}
\description{
Prints the data frame returned by the \code{summary.ordinalNetCV()} method.
}
\examples{
# See ordinalNetCV() documentation for examples.

}
\seealso{
\code{\link{ordinalNetCV}}
}
