% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNetTune.R
\name{print.ordinalNetTune}
\alias{print.ordinalNetTune}
\title{Print method for an "ordinalNetTune" object.}
\usage{
\method{print}{ordinalNetTune}(x, ...)
}
\arguments{
\item{x}{An "ordinalNetTune" S3 object.}

\item{...}{Not used. Additional summary arguments.}
}
\description{
Displays the average out-of-sample log-likelihood and misclassification rate
for each lambda value. The average is taken over all cross validation folds.
}
