% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNet-methods.R
\name{summary.ordinalNet}
\alias{summary.ordinalNet}
\title{Summary method for an "ordinalNet" object.}
\usage{
\method{summary}{ordinalNet}(object, ...)
}
\arguments{
\item{object}{An "ordinalNet" S3 object}

\item{...}{Not used. Additional summary arguments.}
}
\value{
A data frame containing a record for each lambda value in the solution
path. Each record contains the following fields: lambda value, degrees of freedom
(number of nonzero parameters), log-likelihood, AIC, BIC, and percent deviance explained.
}
\description{
Provides a data frame which summarizes the model fit at each lambda value in
the solution path.model fit summary as a data frame.
}
\examples{
# See ordinalNet() documentation for examples.

}
\seealso{
\code{\link{ordinalNet}}
}
