% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNetCV.R
\name{summary.ordinalNetCV}
\alias{summary.ordinalNetCV}
\title{Summary method for an "ordinalNetCV" object.}
\usage{
\method{summary}{ordinalNetCV}(object, ...)
}
\arguments{
\item{object}{An "ordinalNetCV" S3 object}

\item{...}{Not used. Additional summary arguments.}
}
\value{
A data frame containing a record for each cross validation fold.
Each record contains the following: lambda value, log-likelihood,
misclassification rate.
}
\description{
Provides a data frame which summarizes the cross validation results, which
can be used as an estimate of the out-of-sample performance of a model tuned
by a particular method.
}
\examples{
# See ordinalNetCV() documentation for examples.

}
\seealso{
\code{\link{ordinalNetCV}}
}
