% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNetTune.R
\name{summary.ordinalNetTune}
\alias{summary.ordinalNetTune}
\title{Summary method for an "ordinalNetTune" object.}
\usage{
\method{summary}{ordinalNetTune}(object, ...)
}
\arguments{
\item{object}{An "ordinalNetTune" S3 object.}

\item{...}{Not used. Additional summary arguments.}
}
\value{
A data frame containing a record for each lambda value in the solution
path. Each record contains the following: lambda value, average log-likelihood,
average misclassification rate, average Brier score, and average percentage
of deviance explained. Averages are taken across all cross validation folds.
}
\description{
Provides a data frame which summarizes the cross validation results and may
be useful for selecting an appropriate value for the tuning parameter lambda.
}
\examples{
# See ordinalNetTune() documentation for examples.

}
\seealso{
\code{\link{ordinalNetTune}}
}
