% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNetTune.R
\name{plot.ordinalNetTune}
\alias{plot.ordinalNetTune}
\title{Plot method for "ordinalNetTune" object.}
\usage{
\method{plot}{ordinalNetTune}(x, type = c("loglik", "misclass", "brier", "devPct"), ...)
}
\arguments{
\item{x}{An "ordinalNetTune" S3 object.}

\item{type}{Which performance measure to plot. Either "loglik", "misclass",
"brier", or "devPct".}

\item{...}{Additional plot arguments.}
}
\description{
Plots the average out-of-sample log-likelihood, misclassification rate,
Brier score, or percentage of deviance explained for each lambda value in the
solution path. The averae is taken over all cross validation folds.
}
\examples{
# See ordinalNetTune() documentation for examples.

}
\seealso{
\code{\link{ordinalNetTune}}
}
