% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNet-methods.R
\name{print.ordinalNet}
\alias{print.ordinalNet}
\title{Print method for an "ordinalNet" object.}
\usage{
\method{print}{ordinalNet}(x, ...)
}
\arguments{
\item{x}{An "ordinalNet" S3 object}

\item{...}{Not used. Additional plot arguments.}
}
\description{
Prints the data frame returned by the \code{summary.ordinalNet()} method.
}
