% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNetCV.R
\name{print.ordinalNetCV}
\alias{print.ordinalNetCV}
\title{Print method for an "ordinalNetCV" object.}
\usage{
\method{print}{ordinalNetCV}(x, ...)
}
\arguments{
\item{x}{An "ordinalNetCV" S3 object}

\item{...}{Not used. Additional summary arguments.}
}
\description{
Displays the out-of-sample log-likelihood and misclassification rate for each
cross validation fold.
}
