% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNetTune.R
\name{print.ordinalNetTune}
\alias{print.ordinalNetTune}
\title{Print method for an "ordinalNetTune" object.}
\usage{
\method{print}{ordinalNetTune}(x, ...)
}
\arguments{
\item{x}{An "ordinalNetTune" S3 object.}

\item{...}{Not used. Additional print arguments.}
}
\description{
Prints the data frame returned by the \code{summary.ordinalNetTune()} method.
}
\examples{
# See ordinalNetTune() documentation for examples.

}
\seealso{
\code{\link{ordinalNetTune}}
}
